refactor: remove icon on aws-cloud-campaign page and add useToogleMenu custom hook
